FIX: telegram_id BigInteger pour IDs trop grands

+++ database/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime, Boolean, ForeignKey, Text
+from sqlalchemy import Column, Integer, String, Float, DateTime, Boolean, ForeignKey, Text, BigInteger
 from sqlalchemy.orm import relationship
 from .db_manager import Base
 from datetime import datetime  # Ajout de l'import manquant
@@ -6,7 +6,7 @@ from datetime import datetime  # Ajout de l'import manquant
 class User(Base):
     __tablename__ = 'users'
     id = Column(Integer, primary_key=True)
-    telegram_id = Column(Integer, unique=True, nullable=False)
+    telegram_id = Column(BigInteger, unique=True, nullable=False)  # ← CORRECTION CRITIQUE
     username = Column(String(100))
     created_at = Column(DateTime, default=datetime.utcnow, nullable=False)
